packages/agents/pyproject.toml
[project]
name = "pinnacle-agents"
version = "1.0.0"
description = "Pinnacle AI Autonomous Agent Pipeline"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.0",
    "anthropic>=0.40.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "duckdb>=1.2.0",
    "httpx>=0.28.0",
    "pydantic>=2.10.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
]

[tool.setuptools.packages.find]
where = ["."]
